Plot EAF and confidence intervals

This function plots the observed EAF values for each feature in the dataset. The features are ordered by their observed EAF values. The confidence intervals are plotted as error bars or ribbons. The points are colored based on the success ratio of the resamples.

Usage

plot_EAF_values(
  qsip_data_object,
  confidence = 0.9,
  success_ratio = 0.9,
  top = Inf,
  error = "none",
  alpha = 0.3,
  zero_line = TRUE
)

Arguments

qsip_data_object

(qsip_data) A qsip_data object or list of qsip_data objects

confidence

(numeric) The confidence level for the confidence interval

success_ratio

(numeric) The ratio of successful resamples to total resamples

top

(numeric) The number of top features to plot. Use Inf for all

error

(character) The type of error bars to plot. Options are 'none', 'bar', 'ribbon'

alpha

(numeric) The transparency of the error bar/ribbon

zero_line

(logical) Add a line at EAF = 0